VOL-3025 session timeout controller cognito compatibility

See merge request olcs/olcs-selfserve!1228

+++ app/selfserve/module/Olcs/src/Controller/SessionTimeoutController.php
@@ -3,53 +3,38 @@ declare(strict_types=1);
 
 namespace Olcs\Controller;
 
-use Common\Controller\Plugin\CurrentUser;
 use Common\Controller\Plugin\Redirect;
+use Common\Rbac\PidIdentityProvider;
 use Dvsa\Olcs\Auth\Service\Auth\CookieService;
 use Dvsa\Olcs\Auth\Service\Auth\LogoutService;
 use Laminas\Http\Request;
-use Laminas\Http\Response;
 use Laminas\View\Model\ViewModel;
+use ZfcRbac\Identity\IdentityProviderInterface;
 
 /**
  * @See SessionTimeoutControllerFactory
  */
 class SessionTimeoutController
 {
-    /**
-     * @var CurrentUser
-     */
-    private $currentUser;
-
-    /**
-     * @var Redirect
-     */
-    protected $redirectHelper;
-
-    /**
-     * @var LogoutService
-     */
-    private $logoutService;
-
-    /**
-     * @var CookieService
-     */
-    private $cookieService;
+    private IdentityProviderInterface $identityProvider;
+    protected Redirect $redirectHelper;
+    private LogoutService $logoutService;
+    private CookieService $cookieService;
 
     /**
      * SessionTimeoutController constructor.
-     * @param CurrentUser $currentUser
+     * @param IdentityProviderInterface $identityProvider
      * @param Redirect $redirectHelper
      * @param CookieService $cookieService
      * @param LogoutService $logoutService
      */
     public function __construct(
-        CurrentUser $currentUser,
+        IdentityProviderInterface $identityProvider,
         Redirect $redirectHelper,
         CookieService $cookieService,
         LogoutService $logoutService
     ) {
-        $this->currentUser = $currentUser;
+        $this->identityProvider = $identityProvider;
         $this->redirectHelper = $redirectHelper;
         $this->cookieService = $cookieService;
         $this->logoutService = $logoutService;
@@ -62,14 +47,21 @@ class SessionTimeoutController
     public function indexAction(Request $request)
     {
         // redirect to the login
-        $identity = $this->currentUser->getIdentity();
+        $identity = $this->identityProvider->getIdentity();
         if (!is_null($identity) && !$identity->isAnonymous()) {
-            $token = $this->cookieService->getCookie($request);
+            if ($this->identityProvider instanceof PidIdentityProvider) {
+                $token = $this->cookieService->getCookie($request);
 
-            if (!empty($token)) {
-                $this->logoutService->logout($token);
+                if (!empty($token)) {
+                    $this->identityProvider->clearSession();
+                    $this->logoutService->logout($token);
+                    $response = $this->redirectHelper->refresh();
+                    $this->cookieService->destroyCookie($response);
+                    return $response;
+                }
+            } else {
                 $response = $this->redirectHelper->refresh();
-                $this->cookieService->destroyCookie($response);
+                $this->identityProvider->clearSession();
                 return $response;
             }
         }
